refactor(tests): group local corpus suites
